fix: Deposit/Withdraw should use .nextDouble(), not .nextInt()

+++ src/main/MenuOptions/DepositOption.java
@@ -12,7 +12,7 @@ public class DepositOption implements IMenuOption {
         double depositAmount = -1;
         while(depositAmount < 0) {
             System.out.print("How much would you like to deposit: ");
-            depositAmount = ctx.keyboardInput.nextInt();
+            depositAmount = ctx.keyboardInput.nextDouble();
         }
         System.out.println("Successfully Deposited $" + depositAmount);
         ctx.userAccount.deposit(depositAmount);
